+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 
 # Changelog
 
+## `0.2.0` (2/6/2025)
+- Changed API of `ffield` to use node counts instead of bin
+    counts, since it's less confusing
+- Updated docs accordingly
+
 ## `0.1.4` (2/5/2025)
 - Added `fix arbfn/ffield` for more efficient constant force
     fields